Fix the default implementation of IGlobalOperationNotificationService
