+++ src/Workspaces/Core/Portable/Shared/TestHooks/FeatureAttribute_Names.cs
